web/app/requirements.txt
mysql-connector-python
secure-smtplib
weasyprint
Flask
pandas
BeautifulSoup4
requests



